requirements.txt
twikit>=2.0.0
python-dotenv
requests
